Add balanced batching and record limits
batch_data now can yield batches guaranteed to have the
same number of each label (provided they exist in the data).
If this functionality is used, the labels must be hashable
(i.e., not numpy arrays).

It can also be directed to yield only a specific number of records.
Partial batches are not yielded, so the number of records yielded will
be the smaller of max_records or the total number of records in the
dataset, both rounded to batch_size